+++ Sources/SwiftUIReactor/Reactor.swift
@@ -8,20 +8,6 @@
 import Combine
 import SwiftUI
 
-public struct Mutations<Mutation> {
-    public let sync: [Mutation]
-    public let async: AnyPublisher<Mutation, Never>
-    
-    public init(sync: Mutation, async: AnyPublisher<Mutation, Never> = Empty().eraseToAnyPublisher()) {
-        self.init(sync: [sync], async: async)
-    }
-    
-    public init(sync: [Mutation] = [], async: AnyPublisher<Mutation, Never> = Empty().eraseToAnyPublisher()) {
-        self.sync = sync
-        self.async = async
-    }
-}
-
 public protocol Reactor: ObservableObject {
     
     /// An action represents user actions.
@@ -66,6 +52,13 @@ public protocol Reactor: ObservableObject {
 }
 
 public extension Reactor {
+    
+    /// Starts the mutate and reduce chain
+    /// Takes an action, to get all necessary mutations and passes them to the reduce method
+    func action(_ action: Action) {
+        self.action.send(action)
+    }
+    
     func createStateStream() {
         let stateLock = NSLock()
         
@@ -129,115 +122,6 @@ public extension Reactor {
     }
     
     func transform(state: AnyPublisher<State, Never>) -> AnyPublisher<State, Never> {
-        state
-    }
-}
-
-public extension Reactor {
-    
-    /// Starts the mutate and reduce chain
-    /// Takes an action, to get all necessary mutations and passes them to the reduce method
-    func action(_ action: Action) {
-        self.action.send(action)
-    }
-    
-    func mutate<Value>(binding keyPath: KeyPath<State, Value>, _ action: @escaping (Value) -> Action) -> Binding<Value> {
-        Binding<Value>(
-            get: { self.state[keyPath: keyPath] },
-            set: { self.action(action($0)) }
-        )
-    }
-    
-    func reduce<Value>(binding keyPath: KeyPath<State, Value>, _ mutation: @escaping (Value) -> Mutation) -> Binding<Value> {
-        Binding<Value>(
-            get: { self.state[keyPath: keyPath] },
-            set: { self.mutation.send(mutation($0)) }
-        )
-    }
-}
-
-/// Property wrapper to get a binding to a state keyPath and a associated Action
-/// Can be used and behaves like the `@State` property wrapper
-@propertyWrapper
-public struct ActionBinding<R: Reactor, Value>: DynamicProperty {
-    @EnvironmentObject
-    var reactor: R
-    
-    let keyPath: KeyPath<R.State, Value>
-    let action: (Value) -> R.Action
-    
-    /**
-     - Parameters:
-         - reactorType: Type of the reactor in the view´s `EnvironmentObject`
-         - keyPath: Keypath to the value in the reactor´s state
-         - action: Action to perform in the reactor
-     */
-    public init(_ reactorType: R.Type, keyPath: KeyPath<R.State, Value>, action: @escaping (Value) -> R.Action) {
-        self.keyPath = keyPath
-        self.action = action
-    }
-    
-    public var wrappedValue: Value {
-        get { projectedValue.wrappedValue }
-        nonmutating set { projectedValue.wrappedValue = newValue }
-    }
-    
-    public var projectedValue: Binding<Value> {
-        get { reactor.mutate(binding: keyPath, action) }
-    }
-}
-
-/// Property wrapper to get a binding to a state keyPath and a associated Mutation
-/// Can be used and behaves like the `@State` property wrapper
-@propertyWrapper
-public struct MutationBinding<R: Reactor, Value>: DynamicProperty {
-    @EnvironmentObject
-    var reactor: R
-    
-    let keyPath: KeyPath<R.State, Value>
-    let mutation: (Value) -> R.Mutation
-    
-    /**
-     - Parameters:
-         - reactorType: Type of the reactor in the view´s `EnvironmentObject`
-         - keyPath: Keypath to the value in the reactor´s state
-         - mutation: Mutation to perform in the reactor
-     */
-    public init(_ reactorType: R.Type, keyPath: KeyPath<R.State, Value>, mutation: @escaping (Value) -> R.Mutation) {
-        self.keyPath = keyPath
-        self.mutation = mutation
-    }
-    
-    public var wrappedValue: Value {
-        get { projectedValue.wrappedValue }
-        nonmutating set { projectedValue.wrappedValue = newValue }
-    }
-    
-    public var projectedValue: Binding<Value> {
-        get { reactor.reduce(binding: keyPath, mutation) }
-    }
-}
-
-open class BaseReactor<Action, Mutation, State>: Reactor {
-    public let action = PassthroughSubject<Action, Never>()
-    
-    public let mutation = PassthroughSubject<Mutation, Never>()
-    
-    @Published
-    public var state: State
-    
-    public var cancellables = Set<AnyCancellable>()
-    
-    public init(initialState: State) {
-        state = initialState
-        createStateStream()
-    }
-    
-    open func mutate(action: Action) -> Mutations<Mutation> {
-        Mutations()
-    }
-    
-    open func reduce(state: State, mutation: Mutation) -> State {
         state
     }
 }
